use config param to distinguish active parachain light client protocol
