+++ parachain/src/provider.rs
@@ -7,10 +7,7 @@ use ibc::{
 	events::IbcEvent,
 	Height,
 };
-use std::{
-	collections::{BTreeSet, HashMap},
-	fmt::Display,
-};
+use std::fmt::Display;
 
 use ibc_proto::ibc::core::{
 	channel::v1::{
@@ -27,51 +24,33 @@ use sp_runtime::{
 use subxt::Config;
 
 use super::{error::Error, ParachainClient};
-use ibc::core::ics02_client::client_state::ClientState as _;
-use ibc_rpc::{BlockNumberOrHash, IbcApiClient, PacketInfo};
-use primitives::{
-	find_maximum_height_for_timeout_proofs, Chain, IbcProvider, KeyProvider, UpdateType,
-};
+
+use ibc_rpc::{IbcApiClient, PacketInfo};
+use primitives::{Chain, IbcProvider, KeyProvider, UpdateType};
 use sp_core::H256;
 
-use crate::{parachain, parachain::api::runtime_types::primitives::currency::CurrencyId};
+use crate::{
+	parachain, parachain::api::runtime_types::primitives::currency::CurrencyId, GrandpaClientState,
+	LightClientProtocol,
+};
 use ibc::{
 	applications::transfer::{Amount, PrefixedCoin, PrefixedDenom},
-	core::ics02_client::{client_state::ClientType, msgs::update_client::MsgUpdateAnyClient},
+	core::ics02_client::client_state::ClientType,
 	timestamp::Timestamp,
-	tx_msg::Msg,
 };
 use ibc_proto::{
 	google::protobuf::Any,
 	ibc::core::{channel::v1::QueryChannelsResponse, client::v1::IdentifiedClientState},
 };
 
-#[cfg(feature = "beefy")]
-use beefy_light_client_primitives::{ClientState as BeefyPrimitivesClientState, NodesUtils};
+use crate::light_client_protocols::FinalityEvent;
 use beefy_prover::helpers::fetch_timestamp_extrinsic_with_proof;
 use grandpa_light_client_primitives::{FinalityProof, ParachainHeaderProofs};
-#[cfg(feature = "beefy")]
-use ics11_beefy::{
-	client_message::{BeefyHeader, ClientMessage as BeefyClientMessage, ParachainHeadersWithProof},
-	client_state::ClientState as BeefyClientState,
-};
-use pallet_ibc::{
-	light_clients::{AnyClientMessage, AnyClientState, HostFunctionsManager},
-	HostConsensusProof,
-};
-use primitives::mock::LocalClientTypes;
+use ics11_beefy::client_state::ClientState as BeefyClientState;
+use pallet_ibc::{light_clients::HostFunctionsManager, HostConsensusProof};
+
 use sp_runtime::traits::One;
 use std::{collections::BTreeMap, str::FromStr, time::Duration};
-use tendermint_proto::Protobuf;
-
-#[cfg(feature = "beefy")]
-/// Finality event for parachains
-pub type FinalityEvent =
-	beefy_primitives::SignedCommitment<u32, beefy_primitives::crypto::Signature>;
-
-#[cfg(not(feature = "beefy"))]
-pub type FinalityEvent =
-	grandpa_light_client::justification::GrandpaJustification<polkadot_core_primitives::Header>;
 
 #[async_trait::async_trait]
 impl<T: Config + Send + Sync> IbcProvider for ParachainClient<T>
@@ -93,355 +72,30 @@ where
 	type FinalityEvent = FinalityEvent;
 	type Error = Error;
 
-	#[cfg(not(feature = "beefy"))]
 	async fn query_latest_ibc_events<C>(
 		&mut self,
-		justification: Self::FinalityEvent,
+		finality_event: Self::FinalityEvent,
 		counterparty: &C,
 	) -> Result<(Any, Vec<IbcEvent>, UpdateType), anyhow::Error>
 	where
 		C: Chain,
 	{
-		use grandpa_light_client::justification::find_scheduled_change;
-		use grandpa_light_client_primitives::ParachainHeadersWithFinalityProof;
-		use ics10_grandpa::client_message::{ClientMessage, Header as GrandpaHeader};
-
-		let client_id = self.client_id();
-		let latest_height = counterparty.latest_height_and_timestamp().await?.0;
-		let response = counterparty.query_client_state(latest_height, client_id).await?;
-		let client_state = response.client_state.ok_or_else(|| {
-			Error::Custom("Received an empty client state from counterparty".to_string())
-		})?;
-		let client_state = AnyClientState::try_from(client_state)
-			.map_err(|_| Error::Custom("Failed to decode client state".to_string()))?;
-		let grandpa_client_state = match &client_state {
-			AnyClientState::Grandpa(client_state) => client_state,
-			c => Err(Error::ClientStateRehydration(format!(
-				"Expected AnyClientState::Grandpa found: {:?}",
-				c
-			)))?,
-		};
-
-		// fetch the new parachain headers that have been finalized
-		let headers = self
-			.query_finalized_parachain_headers_between(
-				justification.commit.target_hash.into(),
-				grandpa_client_state.latest_relay_hash.into(),
-			)
-			.await?;
-
-		log::info!(
-			"Fetching events from {} for blocks {}..{}",
-			self.name(),
-			headers[0].number(),
-			headers.last().unwrap().number()
-		);
-
-		let finalized_blocks =
-			headers.iter().map(|header| u32::from(*header.number())).collect::<Vec<_>>();
-
-		let finalized_block_numbers = finalized_blocks
-			.iter()
-			.filter_map(|block_number| {
-				if (client_state.latest_height().revision_height as u32) < *block_number {
-					Some(*block_number)
-				} else {
-					None
-				}
-			})
-			.map(|h| BlockNumberOrHash::Number(h))
-			.collect::<Vec<_>>();
-
-		// 1. we should query the sink chain for any outgoing packets to the source chain and return
-		// the maximum height at which we can construct non-existence proofs for all these packets
-		// on the source chain
-		let max_height_for_timeouts =
-			find_maximum_height_for_timeout_proofs(counterparty, self).await;
-		let timeout_update_required = if let Some(max_height) = max_height_for_timeouts {
-			let max_height = max_height as u32;
-			finalized_blocks.contains(&max_height)
-		} else {
-			false
-		};
-
-		let latest_finalized_block = finalized_blocks.into_iter().max().unwrap_or_default();
-
-		let is_update_required = self.is_update_required(
-			latest_finalized_block.into(),
-			client_state.latest_height().revision_height,
-		);
-
-		let target = self
-			.relay_client
-			.rpc()
-			.header(Some(justification.commit.target_hash.into()))
-			.await?
-			.ok_or_else(|| {
-				Error::from(
-					"Could not find relay chain header for justification target".to_string(),
+		match self.light_client_protocol {
+			LightClientProtocol::Grandpa =>
+				LightClientProtocol::query_latest_ibc_event_with_grandpa::<T, C>(
+					self,
+					finality_event,
+					counterparty,
 				)
-			})?;
-
-		let authority_set_changed_scheduled = find_scheduled_change(&target).is_some();
-		// if validator set has changed this is a mandatory update
-		let update_type = match authority_set_changed_scheduled ||
-			timeout_update_required ||
-			is_update_required
-		{
-			true => UpdateType::Mandatory,
-			false => UpdateType::Optional,
-		};
-
-		// block_number => events
-		let events: HashMap<String, Vec<IbcEvent>> = IbcApiClient::<u32, H256>::query_events(
-			&*self.para_client.rpc().client,
-			finalized_block_numbers,
-		)
-		.await?;
-
-		// header number is serialized to string
-		let mut headers_with_events = events
-			.iter()
-			.filter_map(|(num, events)| {
-				if events.is_empty() {
-					None
-				} else {
-					str::parse::<u32>(&*num).ok().map(T::BlockNumber::from)
-				}
-			})
-			.collect::<BTreeSet<_>>();
-
-		let events: Vec<IbcEvent> = events.into_values().flatten().collect();
-
-		if timeout_update_required {
-			let max_height_for_timeouts = max_height_for_timeouts.unwrap();
-			if max_height_for_timeouts > client_state.latest_height().revision_height {
-				let max_timeout_height = T::BlockNumber::from(max_height_for_timeouts as u32);
-				headers_with_events.insert(max_timeout_height);
-			}
-		}
-
-		if is_update_required {
-			headers_with_events.insert(T::BlockNumber::from(latest_finalized_block));
-		}
-
-		let ParachainHeadersWithFinalityProof { finality_proof, parachain_headers } = self
-			.query_grandpa_finalized_parachain_headers_with_proof(
-				justification.commit.target_hash.into(),
-				grandpa_client_state.latest_relay_hash.into(),
-				headers_with_events.into_iter().collect(),
-			)
-			.await?;
-		let grandpa_header = GrandpaHeader {
-			finality_proof: finality_proof.into(),
-			parachain_headers: parachain_headers.into(),
-		};
-
-		for event in events.iter() {
-			if self.sender.send(event.clone()).is_err() {
-				log::trace!("Failed to push {event:?} to stream, no active receiver found");
-				break
-			}
-		}
-
-		let update_header = {
-			let msg = MsgUpdateAnyClient::<LocalClientTypes> {
-				client_id: self.client_id(),
-				client_message: AnyClientMessage::Grandpa(ClientMessage::Header(grandpa_header)),
-				signer: counterparty.account_id(),
-			};
-			let value = msg.encode_vec();
-			Any { value, type_url: msg.type_url() }
-		};
-
-		Ok((update_header, events, update_type))
-	}
-
-	#[cfg(feature = "beefy")]
-	async fn query_latest_ibc_events<C>(
-		&mut self,
-		signed_commitment: Self::FinalityEvent,
-		counterparty: &C,
-	) -> Result<(Any, Vec<IbcEvent>, UpdateType), anyhow::Error>
-	where
-		C: Chain,
-	{
-		let client_id = self.client_id();
-		let latest_height = counterparty.latest_height_and_timestamp().await?.0;
-		let response = counterparty.query_client_state(latest_height, client_id).await?;
-		let client_state = response.client_state.ok_or_else(|| {
-			Error::Custom("Received an empty client state from counterparty".to_string())
-		})?;
-		let client_state = AnyClientState::try_from(client_state)
-			.map_err(|_| Error::Custom("Failed to decode client state".to_string()))?;
-		let beefy_client_state = match &client_state {
-			AnyClientState::Beefy(client_state) => BeefyPrimitivesClientState {
-				latest_beefy_height: client_state.latest_beefy_height,
-				mmr_root_hash: client_state.mmr_root_hash,
-				current_authorities: client_state.authority.clone(),
-				next_authorities: client_state.next_authority_set.clone(),
-				beefy_activation_block: client_state.beefy_activation_block,
-			},
-			c => Err(Error::ClientStateRehydration(format!(
-				"Expected AnyClientState::Beefy found: {:?}",
-				c
-			)))?,
-		};
-
-		if signed_commitment.commitment.validator_set_id < beefy_client_state.current_authorities.id
-		{
-			log::info!(
-				"Commitment: {:#?}\nClientState: {:#?}",
-				signed_commitment.commitment,
-				beefy_client_state
-			);
-			// If validator set id of signed commitment is less than current validator set id we
-			// have Then commitment is outdated and we skip it.
-			log::warn!(
-				"Skipping outdated commitment \n Received signed commitmment with validator_set_id: {:?}\n Current authority set id: {:?}\n Next authority set id: {:?}\n",
-				signed_commitment.commitment.validator_set_id, beefy_client_state.current_authorities.id, beefy_client_state.next_authorities.id
-			);
-			Err(Error::HeaderConstruction("Received an outdated beefy commitment".to_string()))?
-		}
-
-		// fetch the new parachain headers that have been finalized
-		let headers = self
-			.query_finalized_parachain_headers_at(
-				signed_commitment.commitment.block_number,
-				&beefy_client_state,
-			)
-			.await?;
-
-		log::info!(
-			"Fetching events from {} for blocks {}..{}",
-			self.name(),
-			headers[0].number(),
-			headers.last().unwrap().number()
-		);
-
-		// Get finalized parachain block numbers, but only those higher than the latest para
-		// height recorded in the on-chain client state, because in some cases a parachain
-		// block that was already finalized in a former beefy block might still be part of
-		// the parachain headers in a later beefy block, discovered this from previous logs
-		let finalized_blocks =
-			headers.iter().map(|header| u32::from(*header.number())).collect::<Vec<_>>();
-
-		let finalized_block_numbers = finalized_blocks
-			.iter()
-			.filter_map(|block_number| {
-				if (client_state.latest_height().revision_height as u32) < *block_number {
-					Some(*block_number)
-				} else {
-					None
-				}
-			})
-			.map(|h| BlockNumberOrHash::Number(h))
-			.collect::<Vec<_>>();
-
-		// 1. we should query the sink chain for any outgoing packets to the source chain and return
-		// the maximum height at which we can construct non-existence proofs for all these packets
-		// on the source chain
-		let max_height_for_timeouts =
-			find_maximum_height_for_timeout_proofs(counterparty, self).await;
-		let timeout_update_required = if let Some(max_height) = max_height_for_timeouts {
-			let max_height = max_height as u32;
-			finalized_blocks.contains(&max_height)
-		} else {
-			false
-		};
-
-		let latest_finalized_block = finalized_blocks.into_iter().max().unwrap_or_default();
-
-		let authority_set_changed =
-			signed_commitment.commitment.validator_set_id == beefy_client_state.next_authorities.id;
-
-		let is_update_required = self.is_update_required(
-			latest_finalized_block.into(),
-			client_state.latest_height().revision_height,
-		);
-
-		// if validator set has changed this is a mandatory update
-		let update_type =
-			match authority_set_changed || timeout_update_required || is_update_required {
-				true => UpdateType::Mandatory,
-				false => UpdateType::Optional,
-			};
-
-		// block_number => events
-		let events: HashMap<String, Vec<IbcEvent>> = IbcApiClient::<u32, H256>::query_events(
-			&*self.para_client.rpc().client,
-			finalized_block_numbers,
-		)
-		.await?;
-
-		// header number is serialized to string
-		let mut headers_with_events = events
-			.iter()
-			.filter_map(|(num, events)| {
-				if events.is_empty() {
-					None
-				} else {
-					str::parse::<u32>(&*num).ok().map(T::BlockNumber::from)
-				}
-			})
-			.collect::<BTreeSet<_>>();
-
-		let events: Vec<IbcEvent> = events.into_values().flatten().collect();
-
-		if timeout_update_required {
-			let max_height_for_timeouts = max_height_for_timeouts.unwrap();
-			if max_height_for_timeouts > client_state.latest_height().revision_height {
-				let max_timeout_height = T::BlockNumber::from(max_height_for_timeouts as u32);
-				headers_with_events.insert(max_timeout_height);
-			}
-		}
-
-		if is_update_required {
-			headers_with_events.insert(T::BlockNumber::from(latest_finalized_block));
-		}
-
-		// only query proofs for headers that actually have events or are mandatory
-		let headers_with_proof = if !headers_with_events.is_empty() {
-			let (headers, batch_proof) = self
-				.query_beefy_finalized_parachain_headers_with_proof(
-					signed_commitment.commitment.block_number,
-					&beefy_client_state,
-					headers_with_events.into_iter().collect(),
+				.await,
+			LightClientProtocol::Beefy =>
+				LightClientProtocol::query_latest_ibc_event_with_beefy::<T, C>(
+					self,
+					finality_event,
+					counterparty,
 				)
-				.await?;
-			let mmr_size = NodesUtils::new(batch_proof.leaf_count).size();
-
-			Some(ParachainHeadersWithProof {
-				headers,
-				mmr_size,
-				mmr_proofs: batch_proof.items.into_iter().map(|item| item.encode()).collect(),
-			})
-		} else {
-			None
-		};
-
-		let mmr_update =
-			self.fetch_mmr_update_proof_for(signed_commitment, &beefy_client_state).await?;
-		let beefy_header = BeefyHeader { headers_with_proof, mmr_update_proof: Some(mmr_update) };
-
-		for event in events.iter() {
-			if self.sender.send(event.clone()).is_err() {
-				log::trace!("Failed to push {event:?} to stream, no active receiver found");
-				break
-			}
+				.await,
 		}
-
-		let update_header = {
-			let msg = MsgUpdateAnyClient::<LocalClientTypes> {
-				client_id: self.client_id(),
-				client_message: AnyClientMessage::Beefy(BeefyClientMessage::Header(beefy_header)),
-				signer: counterparty.account_id(),
-			};
-			let value = msg.encode_vec();
-			Any { value, type_url: msg.type_url() }
-		};
-
-		Ok((update_header, events, update_type))
 	}
 
 	async fn query_client_consensus(
@@ -801,15 +455,12 @@ where
 		self.client_id()
 	}
 
-	#[cfg(not(feature = "beefy"))]
 	fn client_type(&self) -> ClientType {
-		use ics10_grandpa::client_state::ClientState as GrandpaClientState;
-		GrandpaClientState::<HostFunctionsManager>::client_type()
-	}
-
-	#[cfg(feature = "beefy")]
-	fn client_type(&self) -> ClientType {
-		BeefyClientState::<HostFunctionsManager>::client_type()
+		match self.light_client_protocol {
+			LightClientProtocol::Grandpa =>
+				GrandpaClientState::<HostFunctionsManager>::client_type(),
+			LightClientProtocol::Beefy => BeefyClientState::<HostFunctionsManager>::client_type(),
+		}
 	}
 
 	async fn query_timestamp_at(&self, block_number: u64) -> Result<u64, Self::Error> {
